Let's be a bit more to the point
CTF still exists, I'm still chair, but it's inactive with no plans of me doing anything for it other than handing it over when the time comes. Also, I'm just not good at stating some of these things without being a bit tongue in cheek. The first is really beautiful, but the fifth is something you can really lose yourself in. And I really, really admire the simplicity but freedom of Lisp. If only it had the ecosystem of Python…